% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_database}
\alias{company_database}
\title{Company overviews of 2025 European and American stocks (from Yahoo Finance)}
\format{
A data frame with 2025 rows and 13 columns:
\describe{
  \item{Name}{Company name on one of the Euronext exchanges or Polish brokerage houses}
  \item{Ticker}{Company stock ticker used to collect data from Yahoo Finance}
  \item{Country}{Country name defined with the ISIN of the security}
  \item{Sector}{Company sector of operation defined by Yahoo Finance}
  \item{Industry}{Company industry of operation defined by Yahoo Finance}
  \item{Currency}{Stock quote currency of the particular ticker}
  \item{NoOfEmployees}{Number of company employees in December 2023}
  \item{MarketCap}{Company market capitalization in quote currency units}
  \item{MarketCapUpdate}{Day of the market capitalization sum adequacy}
  \item{TimeSeriesStart}{Beginning date of timeseries in the corresponding `company_timesries` dataset}
  \item{TimeSeriesEnd}{End date of timeseries in the corresponding `company_timesries` dataset}
  \item{EURMarketCap}{Market capitalization in Euros}
  \item{USDMarketCap}{Market capitalization in US Dollars}
}
}
\source{
<https://finance.yahoo.com/quote/>
}
\usage{
company_database
}
\description{
A subset of Yahoo Finance profile and statistics data of 1161 US and
943 European companies (BE, DE, ES, FR, IT, NL, PT)
plus metadata about their stock timeseries in this library
plus their USD and EUR market cap values after conversion
at rate dated at last observation.
}
\keyword{datasets}
