% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_timeseries}
\alias{company_timeseries}
\title{Historical stock prices - monthly closes of 2024 stocks (from Yahoo Finance)}
\format{
An xts object with 71 rows from 2018-01-31 to 2023/11/30 and 2026 columns:
\describe{
  \item{ABCA.PA}{ABC Arbitrage monthly close prices}
  \item{...}{...}
  \item{<Ticker>}{<Company_name> monthly close prices}
  \item{...}{...}
  \item{WMT}{Walmart monthly close prices}
}
}
\source{
<https://finance.yahoo.com/quote/.../history>
}
\usage{
company_timeseries
}
\description{
A subset of Yahoo Finance timeseries history for 2025 stock
monthly close prices adjusted for splits and not for dividends.
Covering the timespan from January 2018 to December 2023.
Column names of this dataset correspond to Yahoo Finance tickers
and the `Ticker`column in the `company_database` dataset.
}
\keyword{datasets}
