% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSectorStructure.R
\name{printSectorStructure}
\alias{printSectorStructure}
\title{Print the nested structure of sectors and industries}
\usage{
printSectorStructure()
}
\value{
[no explicit returns, just a console printout]
}
\description{
`printSectorStructure()` handles a console printout
of the nested strucure of sectors and industries
available in the `company_database` dataset.
}
\details{
For each sector and industry it also prints an example
of a company name and Yahoo ticker on the same line.

Function prints the result slowly to simulate
an actual printout.
}
\examples{
printSectorStructure()

# Author: R. Karol Lisiecki

}
