% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sIndicesChart.R
\name{sIndicesChart}
\alias{sIndicesChart}
\title{Plots aggregate sector market capitalization time series}
\usage{
sIndicesChart(country = NULL, path = NULL)
}
\arguments{
\item{country}{A character vector}

\item{path=NULL}{A character vector}
}
\value{
a plot
}
\description{
sIndicesChart() plots aggregate market capitalizations
for all sectors in a given cross-section
for all months from company_timeseries start point to endpoint
in billions of EUR
using the in-built datasets or data saved by the user.
The optional `path` parameter can point to the directory with saved versions
of the library datasets (with the saveData() function) to use them instead of
the built-in datasets.
}
\examples{

sIndicesChart()  # Chart for all countries combined, using the inbuilt data
sIndicesChart(country = "Italy", path = "./")  # Calculate for Italy using data saved in the current working directory


# Author: R. Karol Lisiecki

}
