% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sIndicesTRCompChart.R
\name{sIndicesTRCompChart}
\alias{sIndicesTRCompChart}
\title{Plots sector total returns comparison time series for two sectors}
\usage{
sIndicesTRCompChart(country1, country2, sector1, sector2, path = NULL)
}
\arguments{
\item{country1}{A character vector}

\item{country2}{A character vector}

\item{sector1}{A character vector}

\item{sector2}{A character vector}

\item{path=NULL}{A character vector}
}
\value{
a plot
}
\description{
sIndicesTRCompChart() plots total returns comparison
of two user-defined sectors in user-defined countries
where the market capitalization in the first period
is treated as the base value (=100).
Return values are calculated in country's base currency
(are not all recalculated to Euros).
}
\details{
The optional `path` parameter can point to the directory with saved versions
of the library datasets (with the saveData() function) to use them instead of
the built-in datasets.
}
\examples{

sIndicesTRCompChart(country1 = "Belgium", country2 = "USA", sector1 = "Technology", sector2 = "Technology")  # Compare the performance of the Technology sector in USA and Belgium

sIndicesTRCompChart(country1 = "France", country2 = "France", sector1 = "Basic Materials", sector2 = "Healthcare")  # Compare the performance of Basic Materials and Healthcare sectors in France

sIndicesTRCompChart(country1 = "Spain", country2 = "Netherlands", sector1 = "Consumer Cyclical", sector2 = "Consumer Defensive")


# Author: R. Karol Lisiecki

}
