% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sStructure.R
\name{sStructure}
\alias{sStructure}
\title{Calculate shares of each sector's market cap. in total market cap.}
\usage{
sStructure(country, path = NULL)
}
\arguments{
\item{country}{A character vector}

\item{path=NULL}{A character vector}
}
\value{
a dataframe
}
\description{
sStructure() calculates each Sector's market capitalization
share in total market capitalization of companies in the dataset,
in the given country, for the most recent market cap. estimate
using the in-built datasets or data saved by the user.
The optional `path` parameter can point to the directory with saved versions
of the library datasets (with the saveData() function) to use them instead of
the built-in datasets.
}
\details{
Function also prints out the total market cap of all companies
in a given country, in billions of EUR.
}
\examples{

sStructure("USA")  # Sector market cap shares of American companies
sStructure(country = "Italy", path = "./")  # Sector market cap shares of Italian companies, calculated from user saved datasets


# Author: R. Karol Lisiecki

}
