% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sStructureChart.R
\name{sStructureChart}
\alias{sStructureChart}
\title{Plots a pie-chart of sector market cap shares for companies in a given country}
\usage{
sStructureChart(country, path = NULL)
}
\arguments{
\item{country}{A character vector}

\item{path=NULL}{A character vector}
}
\value{
a plot
}
\description{
sStructureChart() plots a pie-chart of shares of each sector
in total market capitalization of companies in a given country
using the in-built datasets or data saved by the user.
The optional `path` parameter can point to the directory with saved versions
of the library datasets (with the saveData() function) to use them instead of
the built-in datasets.
}
\details{
The sector labels are augmented with percentatge numbers.
The plot title states the country of investigation.
}
\examples{

sStructureChart("USA")  # Sector market cap shares chart of American companies
sStructureChart(country = "Italy", path = "./")  # Sector market cap shares chart of Italian companies, calculated from user saved datasets


# Author: R. Karol Lisiecki

}
