% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAReport.R
\name{TAReport}
\alias{TAReport}
\title{Calculate a technical analysis report for companies in a given country and sector crossection}
\usage{
TAReport(country, sector, path = NULL)
}
\arguments{
\item{country}{A character vector}

\item{sector}{A character vector}

\item{path=NULL}{A character vector}
}
\value{
a dataframe
}
\description{
TAReport() calculates a technical analysis report
for a defined coutry and sector crossection.
Function assesses the technical sentiment for each company based on two indicators.
First, the Relative Strength Indicator (SMA based 14-period RSI), where
value equal or below 30 is considered oversold,
equal or above 70 overbougt and between those values a no signal zone.
Second, the 11-period Simple Moving Average in relation to the last close price,
where current value above the SMA is considered bullish and below - bearish.
}
\details{
The optional `path` parameter can point to the directory with saved versions
of the library datasets (with the saveData() function) to use them instead of
the built-in datasets.
}
\examples{

TAReport(country = "USA", sector = "Technology") # Technical analysis report for American companies
TAReport(country = "Portugal", sector = "Consumer Defensive", path = "./") # Technical analysis report for Portugese companies, calculated using saved datsets


# Author: R. Karol Lisiecki

}
