% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateData.R
\name{updateData}
\alias{updateData}
\title{Updates the library's datasets saved on disk}
\usage{
updateData(path)
}
\arguments{
\item{path}{A character vector}
}
\description{
updateData() updates company time series,
forex time series, company market capitalization data
and metadata.
}
\details{
Update can take under an hour,
it's status can be observed by watching the progress bar.
User needs to wait until the beginning of the next month to update EOM data.
}
\examples{
# Updates datasets stored in the current working directory
# this is commented because it takes up to fifteen minutes.

# updateData("./")


# Author: R. Karol Lisiecki

}
