% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_timeseries}
\alias{company_timeseries}
\title{Historical stock prices - monthly closes of 2104 stocks (from Yahoo Finance)}
\format{
## `company_timesries`
An xts object on 2018-01-31 / 2023/11/30 containing:
Data: double [71, 2104]
Columns: ABCA.PA, ALCRB.PA, ORIA.PA ... and 2101 more columns (Tickers used in Yahoo Finance)
Index: Date [71] (TZ: "UTC")
}
\source{
<https://finance.yahoo.com/quote/.../history>
}
\usage{
company_timeseries
}
\description{
A subset of Yahoo Finance timeseries history for 2104 stock
monthly close prices adjusted for splits and not for dividends.
Covering the timespan from January 2018 to December 2023.
Column names of this dataset correspond to Yahoo Finance tickers
and the `Ticker`column in the `company_database` dataset.
}
\keyword{datasets}
