% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forex_timeseries}
\alias{forex_timeseries}
\title{Historical FOREX bid prices - monthly close of EURUSD and EURNOK (from Yahoo Finance)}
\format{
An xts object with 71 rows from 2018-01-31 to 2023/11/30 and 2 coumns:
\describe{
  \item{EURUSD}{EUR/USD bid FX spot rate}
  \item{EURNOK}{EUR/NOK bid FX spot rate}
}
}
\source{
<https://finance.yahoo.com/quote/.../history>
}
\usage{
forex_timeseries
}
\description{
A subset of Yahoo Finance timeseries history for EURUSD (`EURUSD=X`)
and EURNOK (`EURNOK=X`) monthly close bid prices.
Covering the timespan from January 2018 to December 2023.
Columns names of this dataset DO NOT correspond to Yahoo Finance tickers
and ARE NOT available in the `company_database` dataset.
}
\keyword{datasets}
