% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Saves libraries data files on user's disk (in Rda format)}
\usage{
saveData(path)
}
\arguments{
\item{path}{A character vector}
}
\description{
saveData() writes all library datasets
inside the folder, for which
the path is given as the function parameter
}
\details{
Enables further data updates
}
\examples{

saveData("./")


# Author: R. Karol Lisiecki

}
